refact() update code

+++ src/main/java/com/techchallenge/msparkingmeter/domain/entity/scheduler/SchedulerInput.java
@@ -11,5 +11,5 @@ public class SchedulerInput {
     private UUID externalDriverId;
     private String phoneNumber;
     private LocalDateTime dateTimeNow;
-    private Long durationInMinutes;
+    private Integer durationInMinutes;
 }
